fix: better ux for duplicate modals
